LayoutTests/fast/css/lang-document-element-flag-expected.txt

Tests that the document element has effectiveLangKnownToMatchDocumentElement flag set to true

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.effectiveLangKnownToMatchDocumentElement(document.documentElement) is true
PASS internals.effectiveLangKnownToMatchDocumentElement(iframeDocElement) is true
PASS internals.effectiveLangKnownToMatchDocumentElement(iframeDocElement) is false
PASS internals.effectiveLangKnownToMatchDocumentElement(body) is true
PASS internals.effectiveLangKnownToMatchDocumentElement(newDocElement) is true
PASS successfullyParsed is true

TEST COMPLETE
